requirements.txt
python-dotenv==1.0.0
fastapi==0.100.0
nltk==3.6.5
requests==2.26.0
beautifulsoup4==4.10.0
uvicorn==0.15.0
validators==0.18.2